fix(): add example
